Moves color generator to utils
