--- src/tokens/color/base.ts
@@ -1,16 +1,0 @@
-import { BaseColorsOptions } from '../../types'
-
-export const baseColorsDefaults: Required<BaseColorsOptions> = {
-  baseColor: '#818181',
-  baseColorIndex: 3,
-  baseColorKey: '100',
-  startLevel: 100,
-  lightdark: 0,
-  saturate: 0,
-  spin: 0,
-  desaturate: 0,
-  greyscale: false,
-  levelsCount: 7,
-  levelGap: 100,
-  tokens: {},
-}
