+++ src/utils/createColorProgressionTokens.ts
@@ -3,9 +3,23 @@ import {
   ColorAdjustment,
   GetColorsOptions,
   Mdfy,
-} from '../../types'
-import { baseColorsDefaults } from './base'
+} from '../types'
 import Color from 'tinycolor2'
+
+const baseColorsDefaults: Required<BaseColorsOptions> = {
+  baseColor: '#818181',
+  baseColorIndex: 3,
+  baseColorKey: '100',
+  startLevel: 100,
+  lightdark: 0,
+  saturate: 0,
+  spin: 0,
+  desaturate: 0,
+  greyscale: false,
+  levelsCount: 7,
+  levelGap: 100,
+  tokens: {},
+}
 
 export const getAdjustment = (
   adjustment: ColorAdjustment,
